refactor: migrate configurePassport and partyServer
