+++ src/http/controller/TablePartyController.ts
@@ -1,10 +1,10 @@
 import { SimpleHandler } from '@/http/HttpHandler'
-import partyWS from '@/modules/internal/partyWebSocket'
+import partyServer from '@/http/websocket/partyServer'
 
 export default class TablePartyController {
   public static upgradeToWebSocket: SimpleHandler = (req, res) => {
-    partyWS.handleUpgrade(req, req.socket, Buffer.from(''), ws => {
-      partyWS.emit('connection', ws, req)
+    partyServer.handleUpgrade(req, req.socket, Buffer.from(''), ws => {
+      partyServer.emit('connection', ws, req)
     })
   }
 }
